refactor(modularization): Phase 5 step 5-S3 — conversation schemas carve (task #26)
Move 21 conversation-owned Pydantic schemas out of
``aperag.schema.view_models`` and into
``aperag/domains/conversation/schemas.py`` using the cuiwenbo-
canonical dual-hook shim from Phase 3 step 4b (``msg=f894cca4`` /
lesson 9a-quad write-only):

* Bot / BotConfig / BotList / BotCreate / BotUpdate / BotUpdateRequest
* Chat / ChatList / ChatCreate / ChatDetails / ChatUpdate / ChatMessage
* Reference / File
* TitleGenerateRequest / TitleGenerateResponse
* TurnFeedback / TurnFeedbackList / Feedback / TurnFeedbackWrite
* Agent  (uses ``aperag.domains.knowledge_base.schemas.Collection`` as
  ``KBCollectionSchema`` — cross-domain import is allowed by G1
  because it does not touch the three legacy aggregate modules)

``view_models.py`` drops the 20 class definitions and gains a new
``try:``-import block at the end, mirroring the KB step 4b block;
when ``view_models`` finishes first the imports succeed, when the
conversation module finishes first its ``_bind_view_models_reexports``
writes the classes back onto the ``view_models`` namespace.
``FailResponse`` stays in ``view_models.py`` — it is not
conversation-specific (marketplace / governance / admin all use it).

Agent-runtime ``AgentMessage`` and the evaluation_v2 Pydantic
request/response shapes stay in place: ``AgentMessage`` moves with
the top-level ``aperag/agent_runtime/`` package rename in 5-S5,
evaluation_v2 schemas move with the top-level ``aperag/evaluation_v2/``
rename in 5-S6. Legacy v1 evaluation schemas (``QuestionSet`` /
``Evaluation`` / ``EvaluationItem`` / etc.) are explicitly **not**
Phase 5 scope and stay in ``view_models`` for Phase 6 cleanup.

Verify:
* ``from aperag.schema.view_models import Bot, Chat, ChatMessage, ...``
  OK — all 21 names resolve
* ``Bot is aperag.domains.conversation.schemas.Bot`` identity preserved
* ``pytest tests/unit_test/`` 671 passed / 29 skipped
* ``uvx ruff check`` + ``uvx ruff format --check`` clean
